requirements.txt
contourpy==1.1.1
cycler==0.12.1
fonttools==4.56.0
guidedfilter==1.0
imageio==2.35.1
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.41.1
matplotlib==3.7.5
mkl-service==2.4.0
networkx==3.1
numba==0.58.1
numpy==1.23.5
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==10.4.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyWavelets==1.4.1
pywin32==306
scikit-image==0.21.0
scipy==1.10.1
tifffile==2023.7.10
torch==1.8.1
torchaudio==0.8.1
torchsummary==1.5.1
torchvision==0.9.1
zipp==3.20.2
